requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
djangorestframework==3.14.0
idna==3.4
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
python-telegram-bot-pagination==0.0.2
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
telebot==0.0.5
tzdata==2023.3
urllib3==2.0.3
